requirements.txt
fastapi
pyjwt
tortoise-orm
uvicorn[standard]
passlib
bcrypt
pyjwt
python-multipart
python-dotenv
pytests
mkdocs
pandas
aerich
polars
pika
tortoise-orm[asyncpg]
mkdocstrings[python]
mkdocs-material
mkdocs-static-i18n